+++ src/module/feature/handlers/commands.handler.ts
@@ -1,4 +1,4 @@
-import { type Client, REST, type RESTPostAPIChatInputApplicationCommandsJSONBody, Routes, type Snowflake } from 'discord.js';
+import { REST, type RESTPostAPIChatInputApplicationCommandsJSONBody, Routes, type Snowflake } from 'discord.js';
 import { readdirSync } from 'fs';
 import { basename } from 'node:path';
 import { join } from 'path';
@@ -8,7 +8,7 @@ import { EmojiEnum } from '../../shared/enums/emoji.enum';
 import { InventorySingleton } from '../../shared/singleton/inventory.singleton';
 import type { SlashCommandModel } from '../slash-commands/model/slash-command.model';
 
-module.exports = async (_client: Client): Promise<void> => {
+module.exports = async (): Promise<void> => {
     const logger: Logger = new Logger(basename(__filename));
     const inventory: InventorySingleton = InventorySingleton.instance;
     const slashCommandsDir: string = join(__dirname, '../slash-commands');
@@ -33,7 +33,7 @@ module.exports = async (_client: Client): Promise<void> => {
         try {
             await rest.put(Routes.applicationGuildCommands(client_id, serverId), { body: command });
 
-            logger.info(`Successfully reloaded application {} slash-commands for guild {}`, String(command.length), serverId);
+            logger.info('Successfully reloaded application {} slash-commands for guild {}', String(command.length), serverId);
         } catch (error) {
             logger.error(`${error}`, error);
         }