+++ src/module/feature/slash-commands/model/watch-clan.model.ts
@@ -92,7 +92,7 @@ export class WatchClanModel {
             return;
         }
 
-        this.logger.info(`Clan {} - {} added to the clan to watch`, id.value as string, name.value as string);
+        this.logger.info('Clan {} - {} added to the clan to watch', id.value as string, name.value as string);
         await interaction.editReply({
             content: 'Le clan a bien été ajouté ! Le clan sera observé à partir du prochain créneaux (*^▽^*)',
         });
@@ -128,7 +128,7 @@ export class WatchClanModel {
             return;
         }
 
-        this.logger.info(`Clan {} - {} removed from the clan to watch`, idOrName.value as string, removed.name);
+        this.logger.info('Clan {} - {} removed from the clan to watch', idOrName.value as string, removed.name);
         await interaction.editReply({ content: 'Le clan a bien été supprimé !' });
 
         this.confirmationEmbed